remount filesystem, skip config mtime check
The remount does fix some occasional fs problems, which seems to just be stale data in fatfs window. This can be invalidated by doing USERFatFS.winsect = -1 but safest to remount properly.

The config ftime still sometimes misses changes if they're in quick succession. This might even be the OS not updating the value, or the low resolution (to the second). But realistically the config file is normally the only thing that gets changed so I don't think always reloading it is a problem.